Cancel spurious @catch force
An optional-widened `@catch` (e.g. `<IO> ?[Int]`) collapses to a
plain-typed value with the `<E>`+`?` migrated onto a CoerceN. Reduce's
`EvalN(DoBlockN)=id` peephole only matched a force sitting directly on a
DoBlockN, so the interposed CoerceN defeated it and codegen emitted
`mlc_catch(...)()` on a value — a C++ compile error and Python `'list'
object is not callable`.

Generalize the native-stage rule to peel the CoerceN stack, then:
commute+cancel a DoBlockN head, drop the spurious force on a
provably-plain head (rebuilding the effect-stripped coerces so `?T`
widening — incl. Rust `Some(..)` — survives), or keep the force on
anything opaque (fail-closed). Runs after checkEffectBoundaries.

+++ test-suite/golden-tests/intrinsic-catch-do-fallback-cpp/helper.hpp
@@ -7,9 +7,11 @@
 
 // Returns a single-element list holding the string length, or throws
 // on empty input so @catch's fallback path is exercised.
-static inline std::vector<int64_t> cpp_maybe_list(const std::string& s) {
+// morloc `Int` maps to C++ `int` (root-cpp: `type Cpp => Int = "int"`), so a
+// `[Int]` return must be `std::vector<int>`.
+static inline std::vector<int> cpp_maybe_list(const std::string& s) {
     if (s.empty()) {
         throw std::runtime_error("cpp_maybe_list: empty input");
     }
-    return std::vector<int64_t>{ static_cast<int64_t>(s.size()) };
+    return std::vector<int>{ static_cast<int>(s.size()) };
 }
